[MJLINK-69] Require Java 11+
