+++ .github/workflows/maven.yml
@@ -25,11 +25,11 @@ jobs:
     strategy:
       matrix:
         os: [ ubuntu-latest, windows-latest, macOS-latest ]
-        java: [ 8, 11, 17 ]
-        dist: [ 'temurin', 'adopt-openj9' ]
+        java: [ 11, 17, 20 ]
+        dist: [ 'temurin', 'semeru' ]
         exclude:
         - os: macos-latest
-          dist: 'adopt-openj9'
+          dist: 'semeru'
 
       fail-fast: false
 
@@ -38,15 +38,6 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v3
-
-# https://github.com/actions/setup-java/blob/main/docs/advanced-usage.md#modifying-maven-toolchains
-      - name: Install Toolchain JDK
-        if: ${{ matrix.java == '8' }}
-        uses: actions/setup-java@v3
-        with:
-          distribution: ${{ matrix.dist }}
-          java-version: 11
-          cache: 'maven'
 
       - name: Set up JDK
         uses: actions/setup-java@v3
