+++ Jenkinsfile
@@ -17,4 +17,6 @@
  * under the License.
  */
 
-asfMavenTlpPlgnBuild()
+asfMavenTlpPlgnBuild([
+  'jdks': [11, 17, 20]
+])
